Segmentation_App/requirements.txt
Pillow
pandas
numpy
matplotlib
medpy
scikit-learn
tensorflow
keras